+++ BancoDeDados/GerenciaConfereIntegridadeJava/ConfereIntegridade/nbproject/private/private.properties
@@ -1,0 +1,7 @@
+compile.on.save=true
+do.depend=false
+do.jar=true
+file.reference.mysql-connector-java-5.1.44-bin.jar=F:\\CEFET-2\u00ba\\Trabalho Final\\mysql-connector-java-5.1.44\\mysql-connector-java-5.1.44-bin.jar
+javac.debug=true
+javadoc.preview=true
+user.properties.file=C:\\Users\\GGzuis\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
