+++ BancoDeDados/GerenciaConfereIntegridadeJava/ConfereIntegridade/nbproject/project.properties
@@ -3,7 +3,7 @@ annotation.processing.enabled.in.editor=false
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=VerificacaoIntegridade
+application.title=ConfereIntegridade
 application.vendor=Aluno
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
@@ -25,15 +25,15 @@ debug.test.classpath=\
 dist.archive.excludes=
 # This directory is removed when the project is cleaned:
 dist.dir=dist
-dist.jar=${dist.dir}/VerificacaoIntegridade.jar
+dist.jar=${dist.dir}/ConfereIntegridade.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
-file.reference.mysql-connector-java-5.1.44-bin.jar=F:\\CEFET-2\u00ba\\Trabalho Final\\mysql-connector-java-5.1.44\\mysql-connector-java-5.1.44-bin.jar
 includes=**
 jar.compress=false
 javac.classpath=\
-    ${file.reference.mysql-connector-java-5.1.44-bin.jar}
+    ${file.reference.mysql-connector-java-5.1.44-bin.jar}:\
+    ${libs.MySQLDriver.classpath}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
@@ -58,7 +58,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=VerificacaoIntegridade
+main.class=TesteConfereIntegridade
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
